Add per-lesson reset to the Lesson pannel
Deduplicates the code in the Sidepanel classes of LessonsPanel.py
and PuzzlesPanel.py, by putting it in a function, in it's own file:
generate/generateLessonsSidepanel.py.
